studio-2975906357:~/studio/functions{master}$ npm run lint
> lint
> ./node_modules/.bin/eslint --config .eslintrc.js --ext .js,.ts .

=============

WARNING: You are currently running a version of TypeScript which is not officially supported by @typescript-eslint/typescript-estree.

You may find that it works just fine, or you may not.

SUPPORTED TYPESCRIPT VERSIONS: >=3.3.1 <5.2.0

YOUR TYPESCRIPT VERSION: 5.8.3

Please only submit bug reports when using the officially supported version.

=============

/home/user/studio/functions/src/processEmployeeCSV.ts
  131:48  warning  Unexpected any. Specify a different type               @typescript-eslint/no-explicit-any
  144:1   error    This line has a length of 123. Maximum allowed is 120  max-len
  231:19  error    Strings must use doublequote                           quotes
  263:19  error    Strings must use doublequote                           quotes

✖ 4 problems (3 errors, 1 warning)
  2 errors and 0 warnings potentially fixable with the `--fix` option.

Why there are so many issues. Fix all the issues once and for all. If required, you can take alternative methods to make this work. This csv file data transfer is a one time feature for me. After I upload the data of 333 guards i will be using this website to enroll guards. So if there is any easy method you can adopt that.